Vérification des commentaires et des de la présence de codes commentés

+++ DotNetEnglishP3-master/P3AddNewFunctionalityDotNetCore/Models/Services/ProductService.cs
@@ -153,11 +153,15 @@ namespace P3AddNewFunctionalityDotNetCore.Models.Services
         public void DeleteProduct(int id)
         {
             //When this method is used it delete the product everywhere in the list and in the cart if the product was added in
-            
-
-            _cart.RemoveLine(GetProductById(id));
-
-            _productRepository.DeleteProduct(id);
+            var product = _productRepository.GetProduct(id).Result;
+            if (product != null)
+            {
+                _productRepository.DeleteProduct(id);
+            }
+            else
+            {
+                throw new InvalidOperationException($"Le produit avec l'ID {id} n'existe pas.");
+            }
         }
     }
 }
